Move the `CredentialSource` class definitions into the `.cc` file
None of these are used directly as part of the public API.
