+++ base/cvd/cuttlefish/host/libs/web/credential_source.cc
@@ -41,7 +41,85 @@
 namespace cuttlefish {
 namespace {
 
+constexpr char kBuildScope[] =
+    "https://www.googleapis.com/auth/androidbuild.internal";
+
 constexpr auto kRefreshWindow = std::chrono::minutes(2);
+
+class GceMetadataCredentialSource : public CredentialSource {
+ public:
+  GceMetadataCredentialSource(HttpClient&);
+  GceMetadataCredentialSource(GceMetadataCredentialSource&&) = default;
+
+  Result<std::string> Credential() override;
+
+  static std::unique_ptr<CredentialSource> Make(HttpClient&);
+
+ private:
+  HttpClient& http_client_;
+  std::string latest_credential_;
+  std::chrono::steady_clock::time_point expiration_;
+
+  Result<void> RefreshCredential();
+};
+
+class FixedCredentialSource : public CredentialSource {
+ public:
+  FixedCredentialSource(const std::string& credential);
+
+  Result<std::string> Credential() override;
+
+  static std::unique_ptr<CredentialSource> Make(const std::string& credential);
+
+ private:
+  std::string credential_;
+};
+
+class RefreshCredentialSource : public CredentialSource {
+ public:
+  static Result<RefreshCredentialSource> FromOauth2ClientFile(
+      HttpClient& http_client, const std::string& oauthcontents);
+
+  RefreshCredentialSource(HttpClient& http_client, const std::string& client_id,
+                          const std::string& client_secret,
+                          const std::string& refresh_token);
+
+  Result<std::string> Credential() override;
+
+ private:
+  Result<void> UpdateLatestCredential();
+
+  HttpClient& http_client_;
+  std::string client_id_;
+  std::string client_secret_;
+  std::string refresh_token_;
+
+  std::string latest_credential_;
+  std::chrono::steady_clock::time_point expiration_;
+};
+
+class ServiceAccountOauthCredentialSource : public CredentialSource {
+ public:
+  static Result<ServiceAccountOauthCredentialSource> FromJson(
+      HttpClient& http_client, const Json::Value& service_account_json,
+      const std::string& scope);
+  ServiceAccountOauthCredentialSource(ServiceAccountOauthCredentialSource&&) =
+      default;
+
+  Result<std::string> Credential() override;
+
+ private:
+  ServiceAccountOauthCredentialSource(HttpClient& http_client);
+  Result<void> RefreshCredential();
+
+  HttpClient& http_client_;
+  std::string email_;
+  std::string scope_;
+  std::unique_ptr<EVP_PKEY, void (*)(EVP_PKEY*)> private_key_;
+
+  std::string latest_credential_;
+  std::chrono::steady_clock::time_point expiration_;
+};
 
 std::unique_ptr<CredentialSource> TryParseServiceAccount(
     HttpClient& http_client, const std::string& file_content) {
@@ -107,8 +185,6 @@ Result<std::unique_ptr<CredentialSource>> GetCredentialSourceLegacy(
   }
   return result;
 }
-
-}  // namespace
 
 GceMetadataCredentialSource::GceMetadataCredentialSource(
     HttpClient& http_client)
@@ -430,6 +506,8 @@ Result<std::string> ServiceAccountOauthCredentialSource::Credential() {
   }
   return latest_credential_;
 }
+
+}  // namespace
 
 Result<std::unique_ptr<CredentialSource>> GetCredentialSource(
     HttpClient& http_client, const std::string& credential_source,
