+++ base/cvd/cuttlefish/host/libs/web/credential_source.h
@@ -15,99 +15,18 @@
 
 #pragma once
 
-#include <chrono>
 #include <memory>
 #include <string>
-
-#include "openssl/evp.h"
 
 #include "common/libs/utils/result.h"
 #include "host/libs/web/http_client/http_client.h"
 
 namespace cuttlefish {
 
-inline constexpr char kBuildScope[] =
-    "https://www.googleapis.com/auth/androidbuild.internal";
-
 class CredentialSource {
  public:
   virtual ~CredentialSource() = default;
   virtual Result<std::string> Credential() = 0;
-};
-
-class GceMetadataCredentialSource : public CredentialSource {
- public:
-  GceMetadataCredentialSource(HttpClient&);
-  GceMetadataCredentialSource(GceMetadataCredentialSource&&) = default;
-
-  Result<std::string> Credential() override;
-
-  static std::unique_ptr<CredentialSource> Make(HttpClient&);
-
- private:
-  HttpClient& http_client_;
-  std::string latest_credential_;
-  std::chrono::steady_clock::time_point expiration_;
-
-  Result<void> RefreshCredential();
-};
-
-class FixedCredentialSource : public CredentialSource {
- public:
-  FixedCredentialSource(const std::string& credential);
-
-  Result<std::string> Credential() override;
-
-  static std::unique_ptr<CredentialSource> Make(const std::string& credential);
-
- private:
-  std::string credential_;
-};
-
-class RefreshCredentialSource : public CredentialSource {
- public:
-  static Result<RefreshCredentialSource> FromOauth2ClientFile(
-      HttpClient& http_client, const std::string& oauthcontents);
-
-  RefreshCredentialSource(HttpClient& http_client, const std::string& client_id,
-                          const std::string& client_secret,
-                          const std::string& refresh_token);
-
-  Result<std::string> Credential() override;
-
- private:
-  Result<void> UpdateLatestCredential();
-
-  HttpClient& http_client_;
-  std::string client_id_;
-  std::string client_secret_;
-  std::string refresh_token_;
-
-  std::string latest_credential_;
-  std::chrono::steady_clock::time_point expiration_;
-};
-
-class ServiceAccountOauthCredentialSource : public CredentialSource {
- public:
-  static Result<ServiceAccountOauthCredentialSource> FromJson(
-      HttpClient& http_client, const Json::Value& service_account_json,
-      const std::string& scope);
-  ServiceAccountOauthCredentialSource(ServiceAccountOauthCredentialSource&&) =
-      default;
-
-  Result<std::string> Credential() override;
-
- private:
-  ServiceAccountOauthCredentialSource(HttpClient& http_client);
-  Result<void> RefreshCredential();
-
-  HttpClient& http_client_;
-  std::string email_;
-  std::string scope_;
-  std::unique_ptr<EVP_PKEY, void (*)(EVP_PKEY*)> private_key_;
-
-  std::string latest_credential_;
-  std::chrono::steady_clock::time_point expiration_;
 };
 
 Result<std::unique_ptr<CredentialSource>> GetCredentialSource(
